+++ fearless-utils/src/test/java/jp/co/soramitsu/fearless_utils/common/NetworkTypes.kt
@@ -1,9 +1,9 @@
 package jp.co.soramitsu.fearless_utils.common
 
 internal object TestAddressBytes {
-    const val KUSAMA: Byte = 2
-    const val POLKADOT: Byte = 0
-    const val WESTEND: Byte = 42
+    const val KUSAMA: Short = 2
+    const val POLKADOT: Short = 0
+    const val WESTEND: Short = 42
 }
 
 internal object TestGeneses {
